add support for proto2

+++ src/google/protobuf/map_field_inl.h
@@ -168,7 +168,7 @@ template <typename Derived, typename Key, typename T,
 int MapField<Derived, Key, T, kKeyFieldType, kValueFieldType,
              default_enum_value>::size() const {
   MapFieldBase::SyncMapWithRepeatedField();
-  return static_cast<int>(impl_.GetMap().size());
+  return impl_.GetMap().size();
 }
 
 template <typename Derived, typename Key, typename T,
